[Refactor][Adrian]
