+++ src/components/component/student/enrollment/header.jsx
@@ -1,0 +1,11 @@
+function EnrollmentHeader() {
+  return (
+    <div className='mt-32 flex flex-col place-items-start'>
+      <h1 className='font-medium text-4xl text-left mb-2'>Enrollment</h1>
+      <p className='text-zinc-400'>Current School Year / Term</p>
+      <p className='mb-10'>School Year 2023 - 2024 1st Trimester</p>
+    </div>
+  );
+}
+
+export default EnrollmentHeader;
